Bind CursorMoved event only to the current buffer
Previously the CursorMoved event used to auto-close the popup window
was bounded globally, so when a cursor is to moved in another window
or even when the popup window gets a focus, the popup window would be
closed. This makes impossible to move the cursor to the popup window.

With the event bounded only to the current buffer, we can fix this problem.